Check of any field deleted

+++ src/main/java/seedu/address/logic/parser/DeleteFieldCommandParser.java
@@ -159,6 +159,10 @@ public class DeleteFieldCommandParser implements Parser<DeleteFieldCommand> {
             }
         }
 
+        if (!deleteFieldDescriptor.isAnyFieldDeleted()) {
+            throw new ParseException(DeleteFieldCommand.MESSAGE_NOT_DELETED_FIELD);
+        }
+
         return new DeleteFieldCommand(deleteFieldDescriptor, personToDeleteField);
     }
 }
